Add new layout files for friend selection, recommendations, social activity, and user search
- Created `item_friend_select.xml` for displaying friends with avatars and checkboxes.
- Added `item_recommendation.xml` for showing movie recommendations with user info and actions.
- Introduced `item_social_activity.xml` to display user activities related to movies.
- Implemented `item_user_search.xml` for searching users with friend request options.

Enhance menu and string resources for friend options and social features

- Added `menu_friend_options.xml` for friend-related actions.
- Updated `strings.xml` for English and Russian languages with new strings for social features, friend requests, and recommendations.

Update color resources and styles for UI consistency

- Added new colors for genre chips in `colors.xml`.
- Defined a new style for tab text in `styles.xml`.
- Created a transparent dialog style in `themes.xml`.

Update build configuration for Google services

- Added Google services plugin in `build.gradle.kts`.

+++ app/src/main/java/com/example/movietime/data/repository/AppRepository.kt
@@ -531,7 +531,7 @@ class AppRepository @Inject constructor(
             posterPath = movie.posterPath,
             mediaType = "movie",
             releaseDate = movie.releaseDate,
-            voteAverage = movie.voteAverage
+            voteAverage = movie.voteAverage.toDouble()
         )
         searchHistoryDao.insert(historyItem)
     }
@@ -543,7 +543,7 @@ class AppRepository @Inject constructor(
             posterPath = tvShow.posterPath,
             mediaType = "tv",
             releaseDate = tvShow.firstAirDate,
-            voteAverage = tvShow.voteAverage
+            voteAverage = tvShow.voteAverage.toDouble()
         )
         searchHistoryDao.insert(historyItem)
     }
